Fix: remove unnecessary nullsafe operators and fix relationships
- Remove unnecessary ?-> operators where objects are never null
- Fix EquipmentItem relationship usage in ApplicationForm (use category instead of equipment)
- Cast float to int in EquipmentLoan::getLoanDurationAttribute()
- Remove unused closure variables to eliminate warnings

+++ app/Models/EquipmentLoan.php
@@ -64,7 +64,7 @@ class EquipmentLoan extends Model
      */
     public function getLoanDurationAttribute(): int
     {
-        return Carbon::parse($this->loan_start_date)->diffInDays(Carbon::parse($this->loan_end_date)) + 1;
+        return (int) (Carbon::parse($this->loan_start_date)->diffInDays(Carbon::parse($this->loan_end_date)) + 1);
     }
 
     /**
